component1:Rack Description: 1.This Component will render the bookshelf along with designated books 2.there are 3 props passed along with:bookshehtitle prop/bookshelfname prop/Books prop 3.This same  component will render three time depending on their different props that has passed on.
